+++ Projects/NUCLEO-F401RE/Examples/IKS01A3/LSM6DSO_FIFOMode/Src/app_mems.c
@@ -1,21 +1,20 @@
 /**
- ******************************************************************************
- * File Name          : app_mems.c
- * Description        : This file provides code for the configuration
- *                      of the STMicroelectronics.X-CUBE-MEMS1.9.2.0 instances.
- ******************************************************************************
- * @attention
- *
- * <h2><center>&copy; Copyright (c) 2022 STMicroelectronics.
- * All rights reserved.</center></h2>
- *
- * This software component is licensed by ST under Software License Agreement SLA0077,
- * the "License". You may not use this component except in compliance with the
- * License. You may obtain a copy of the License at:
- *                        www.st.com/sla0077
- *
- ******************************************************************************
- */
+  ******************************************************************************
+  * File Name          : app_mems.c
+  * Description        : This file provides code for the configuration
+  *                      of the STMicroelectronics.X-CUBE-MEMS1.9.3.0 instances.
+  ******************************************************************************
+  * @attention
+  *
+  * Copyright (c) 2022 STMicroelectronics.
+  * All rights reserved.
+  *
+  * This software is licensed under terms that can be found in the LICENSE file
+  * in the root directory of this software component.
+  * If no LICENSE file comes with this software, it is provided AS-IS.
+  *
+  ******************************************************************************
+  */
 
 #ifdef __cplusplus
 extern "C" {
@@ -392,7 +391,7 @@ static int32_t LSM6DSO_Read_Single_FIFO_Data(uint16_t SampleIndex)
     return ret;
   }
 
-  (void)snprintf(dataOut, MAX_BUF_SIZE, "[DATA %02d]  %8ld  %8ld  %8ld\r\n", SampleIndex + 1U, angular_velocity.x, angular_velocity.y, angular_velocity.z);
+  (void)snprintf(dataOut, MAX_BUF_SIZE, "[DATA %02d]  %8ld  %8ld  %8ld\r\n", SampleIndex + 1U, (long)angular_velocity.x, (long)angular_velocity.y, (long)angular_velocity.z);
   printf("%s", dataOut);
 
   return ret;
@@ -402,4 +401,3 @@ static int32_t LSM6DSO_Read_Single_FIFO_Data(uint16_t SampleIndex)
 }
 #endif
 
-/************************ (C) COPYRIGHT STMicroelectronics *****END OF FILE****/
